feat: add create company
